requirements.txt
flask
gitpython